Warn and prompt the user when a book file needs ot be overwritten
Generated book files will remain the unless the user deletes them.
Since Roam appends blocks when importing Markdown files, it is
better to only write new highlights to files. For this reason,
prompt the user whether the file should be overwritten or not.

For now, this will advance the last updated time regardless if
it is overwritten or not, so new highlights will be lost if they
are not accepted.